Abstractify performance points calculation to a base class
