+++ osu.Game/Screens/Play/HUD/PerformancePointsCounter.cs
@@ -13,16 +13,9 @@ using JetBrains.Annotations;
 using osu.Framework.Allocation;
 using osu.Framework.Audio.Track;
 using osu.Framework.Extensions;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
-using osu.Framework.Localisation;
 using osu.Game.Beatmaps;
-using osu.Game.Graphics;
-using osu.Game.Graphics.Sprites;
 using osu.Game.Graphics.UserInterface;
-using osu.Game.Resources.Localisation.Web;
 using osu.Game.Rulesets;
 using osu.Game.Rulesets.Difficulty;
 using osu.Game.Rulesets.Judgements;
@@ -31,19 +24,12 @@ using osu.Game.Rulesets.Objects;
 using osu.Game.Rulesets.Scoring;
 using osu.Game.Scoring;
 using osu.Game.Skinning;
-using osuTK;
 
 namespace osu.Game.Screens.Play.HUD
 {
     public partial class PerformancePointsCounter : RollingCounter<int>, ISerialisableDrawable
     {
         public bool UsesFixedAnchor { get; set; }
-
-        protected override bool IsRollingProportional => true;
-
-        protected override double RollingDuration => 500;
-
-        private const float alpha_when_invalid = 0.3f;
 
         [Resolved]
         private ScoreProcessor scoreProcessor { get; set; }
@@ -60,18 +46,11 @@ namespace osu.Game.Screens.Play.HUD
         private PerformanceCalculator performanceCalculator;
         private ScoreInfo scoreInfo;
 
-        public PerformancePointsCounter()
-        {
-            Current.Value = DisplayedCount = 0;
-        }
-
         private Mod[] clonedMods;
 
         [BackgroundDependencyLoader]
-        private void load(OsuColour colours, BeatmapDifficultyCache difficultyCache)
+        private void load(BeatmapDifficultyCache difficultyCache)
         {
-            Colour = colours.BlueLighter;
-
             if (gameplayState != null)
             {
                 performanceCalculator = gameplayState.Ruleset.CreatePerformanceCalculator();
@@ -107,19 +86,7 @@ namespace osu.Game.Screens.Play.HUD
                 onJudgementChanged(gameplayState.LastJudgementResult.Value);
         }
 
-        private bool isValid;
-
-        protected bool IsValid
-        {
-            set
-            {
-                if (value == isValid)
-                    return;
-
-                isValid = value;
-                DrawableCount.FadeTo(isValid ? 1 : alpha_when_invalid, 1000, Easing.OutQuint);
-            }
-        }
+        public virtual bool IsValid { get; set; }
 
         private void onJudgementChanged(JudgementResult judgement)
         {
@@ -151,13 +118,6 @@ namespace osu.Game.Screens.Play.HUD
             return timedAttributes[Math.Clamp(attribIndex, 0, timedAttributes.Count - 1)].Attributes;
         }
 
-        protected override LocalisableString FormatCount(int count) => count.ToString(@"D");
-
-        protected override IHasText CreateText() => new TextComponent
-        {
-            Alpha = alpha_when_invalid
-        };
-
         protected override void Dispose(bool isDisposing)
         {
             base.Dispose(isDisposing);
@@ -169,45 +129,6 @@ namespace osu.Game.Screens.Play.HUD
             }
 
             loadCancellationSource?.Cancel();
-        }
-
-        private partial class TextComponent : CompositeDrawable, IHasText
-        {
-            public LocalisableString Text
-            {
-                get => text.Text;
-                set => text.Text = value;
-            }
-
-            private readonly OsuSpriteText text;
-
-            public TextComponent()
-            {
-                AutoSizeAxes = Axes.Both;
-
-                InternalChild = new FillFlowContainer
-                {
-                    AutoSizeAxes = Axes.Both,
-                    Spacing = new Vector2(2),
-                    Children = new Drawable[]
-                    {
-                        text = new OsuSpriteText
-                        {
-                            Anchor = Anchor.BottomLeft,
-                            Origin = Anchor.BottomLeft,
-                            Font = OsuFont.Numeric.With(size: 16, fixedWidth: true)
-                        },
-                        new OsuSpriteText
-                        {
-                            Anchor = Anchor.BottomLeft,
-                            Origin = Anchor.BottomLeft,
-                            Text = BeatmapsetsStrings.ShowScoreboardHeaderspp,
-                            Font = OsuFont.Numeric.With(size: 8),
-                            Padding = new MarginPadding { Bottom = 1.5f }, // align baseline better
-                        }
-                    }
-                };
-            }
         }
 
         // TODO: This class shouldn't exist, but requires breaking changes to allow DifficultyCalculator to receive an IBeatmap.
